+++ models/Usuarios.php
@@ -275,9 +275,11 @@ class Usuarios extends \yii\db\ActiveRecord implements \yii\web\IdentityInterfac
     {
         if ($this->imagen !== null) {
             try {
-                $img = Utility::s3Download($this->imagen, 'netflox-usuarios');
                 $path = Yii::getAlias('@user/' . $this->imagen);
-                file_put_contents($path, $img['Body']);
+                if (!file_exists($path)) {
+                    $img = Utility::s3Download($this->imagen, 'netflox-usuarios');
+                    file_put_contents($path, $img['Body']);
+                }
                 return $path;
             } catch (\Exception $exception) {
             }
